// src/app/models/etiqueta.model.ts
export class  EyeEtiqueta {
    c_codigo_eti: string;
    v_nombre_eti: string;
    v_imagen_eti:  string;
    c_codigoalt_eti: string;
    v_nombreingles_eti: string;
    c_clave_eti: string;
    v_abreviatura_eti: string;
    c_texper_eti: string;
    v_marca_eti: string;


    constructor() {
        this.c_codigo_eti = '';
        this.v_nombre_eti= '';
        this.v_imagen_eti= '';
        this.c_codigoalt_eti= '';
        this.v_nombreingles_eti= '';
        this.c_clave_eti= '';
        this.v_abreviatura_eti= '';
        this.c_texper_eti= '';
        this.v_marca_eti= '';
    
    }
  }